+++ src/layouts/RootLayout/Scripts.tsx
