+++ src/layouts/RootLayout/index.tsx
@@ -1,6 +1,9 @@
 import React, { ReactNode } from "react"
 import { ThemeProvider } from "./ThemeProvider"
 import useScheme from "src/hooks/useScheme"
+import Header from "./Header"
+import styled from "@emotion/styled"
+import Scripts from "src/layouts/RootLayout/Scripts"
 
 type Props = {
   children: ReactNode
@@ -8,7 +11,22 @@ type Props = {
 
 const RootLayout = ({ children }: Props) => {
   const [scheme] = useScheme()
-  return <ThemeProvider scheme={scheme}>{children}</ThemeProvider>
+  return (
+    <ThemeProvider scheme={scheme}>
+      <Scripts />
+      {/* // TODO: replace react query */}
+      {/* {metaConfig.type !== "Paper" && <Header />} */}
+      <Header fullWidth={false} />
+      <StyledMain>{children}</StyledMain>
+    </ThemeProvider>
+  )
 }
 
 export default RootLayout
+
+const StyledMain = styled.main`
+  margin: 0 auto;
+  width: 100%;
+  max-width: 1120px;
+  padding: 0 1rem;
+`
