+++ src/pages/_app.tsx
@@ -1,21 +1,19 @@
 import useGtagEffect from "src/hooks/useGtagEffect"
-import Scripts from "src/components/Scripts"
 import { AppPropsWithLayout } from "../types"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
-import RootLayout from "src/layouts/RootLayout"
+import { RootLayout } from "src/layouts"
 
 const queryClient = new QueryClient()
 
-function MyApp({ Component, pageProps }: AppPropsWithLayout) {
+function App({ Component, pageProps }: AppPropsWithLayout) {
   const getLayout = Component.getLayout || ((page) => page)
   useGtagEffect()
 
   return (
     <QueryClientProvider client={queryClient}>
-      <Scripts />
       <RootLayout>{getLayout(<Component {...pageProps} />)}</RootLayout>
     </QueryClientProvider>
   )
 }
 
-export default MyApp
+export default App
